requirements.txt
# ----------------------
# Flask
# ----------------------
flask
Flask-WTF
wtforms
Flask-SQLAlchemy
validators

# ----------------------
# Testing
# ----------------------
mock


# ----------------------
# Production Server
# ----------------------
gunicorn


